+++ src/main/java/com/account/yomankum/security/domain/JwtValue.java
@@ -1,4 +1,4 @@
-package com.account.yomankum.security.oauth;
+package com.account.yomankum.security.domain;
 
 public interface JwtValue {
