+++ src/main/java/com/account/yomankum/security/domain/KakaoUserInfo.java
@@ -1,4 +1,7 @@
-package com.account.yomankum.security.oauth;
+package com.account.yomankum.security.domain;
+
+import com.account.yomankum.security.domain.Sns;
+import com.account.yomankum.security.domain.SnsUserInfo;
 
 import java.util.Map;
